+++ README.md
@@ -3,13 +3,24 @@
 [![Build Status](https://secure.travis-ci.org/Kronos-Integration/kronos-service-uti.png)](http://travis-ci.org/Kronos-Integration/kronos-service-uti)
 [![bithound](https://www.bithound.io/github/Kronos-Integration/kronos-service-uti/badges/score.svg)](https://www.bithound.io/github/Kronos-Integration/kronos-service-uti)
 [![codecov.io](http://codecov.io/github/Kronos-Integration/kronos-service-uti/coverage.svg?branch=master)](http://codecov.io/github/Kronos-Integration/kronos-service-uti?branch=master)
+[![Coverage Status](https://coveralls.io/repos/Kronos-Integration/kronos-service-uti/badge.svg)](https://coveralls.io/r/Kronos-Integration/kronos-service-uti)
 [![Code Climate](https://codeclimate.com/github/Kronos-Integration/kronos-service-uti/badges/gpa.svg)](https://codeclimate.com/github/Kronos-Integration/kronos-service-uti)
+[![Known Vulnerabilities](https://snyk.io/test/github/Kronos-Integration/kronos-service-uti/badge.svg)](https://snyk.io/test/github/Kronos-Integration/kronos-service-uti)
 [![GitHub Issues](https://img.shields.io/github/issues/Kronos-Integration/kronos-service-uti.svg?style=flat-square)](https://github.com/Kronos-Integration/kronos-service-uti/issues)
+[![Stories in Ready](https://badge.waffle.io/Kronos-Integration/kronos-service-uti.svg?label=ready&title=Ready)](http://waffle.io/Kronos-Integration/kronos-service-uti)
 [![Dependency Status](https://david-dm.org/Kronos-Integration/kronos-service-uti.svg)](https://david-dm.org/Kronos-Integration/kronos-service-uti)
 [![devDependency Status](https://david-dm.org/Kronos-Integration/kronos-service-uti/dev-status.svg)](https://david-dm.org/Kronos-Integration/kronos-service-uti#info=devDependencies)
 [![docs](http://inch-ci.org/github/Kronos-Integration/kronos-service-uti.svg?branch=master)](http://inch-ci.org/github/Kronos-Integration/kronos-service-uti)
 [![downloads](http://img.shields.io/npm/dm/kronos-service-uti.svg?style=flat-square)](https://npmjs.org/package/kronos-service-uti)
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
+
+kronos-service-uti
+=====
+uti provider service
+
+# API Reference
+
+* * *
 
 install
 =======
