+++ doc/README.hbs
@@ -14,3 +14,29 @@
 [![downloads](http://img.shields.io/npm/dm/kronos-service-uti.svg?style=flat-square)](https://npmjs.org/package/kronos-service-uti)
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
 
+kronos-service-uti
+=====
+uti provider service
+
+# API Reference
+{{#modules~}}
+- {{name}}
+{{/modules}}
+
+{{#functions}}* {{>docs}}
+{{/functions}}
+* * *
+
+install
+=======
+
+With [npm](http://npmjs.org) do:
+
+```shell
+npm install kronos-service-uti
+```
+
+license
+=======
+
+BSD-2-Clause
