--- doc/index.adoc
@@ -1,2 +1,0 @@
-== {name}
-any doc
