+++ package.json
@@ -5,13 +5,13 @@
   "keywords": [
     "kronos-service"
   ],
-  "main": "service.js",
+  "module": "src/ServiceUTI.js",
+  "main": "dist/module.js",
   "scripts": {
     "cover": "npm run build && node_modules/.bin/istanbul cover node_modules/.bin/_mocha -- -u exports tests/*_test.js",
-    "doc": "./node_modules/.bin/jsdoc lib/*.js",
     "test": "npm run build && node_modules/.bin/mocha tests/*_test.js && markdown-doctest",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
-    "build": "node_modules/rollup/bin/rollup --output=service.js -c -- {{module}}",
+    "build": "node_modules/rollup/bin/rollup --output=dist/module.js -c -- src/ServiceUTI.js",
     "docs": "jsdoc2md -l off -t doc/README.hbs -f src/*.js > README.md"
   },
   "repository": {
@@ -23,14 +23,13 @@
   },
   "homepage": "https://github.com/Kronos-Integration/kronos-service-uti#readme",
   "dependencies": {
-    "kronos-service": "^4.1.6",
+    "kronos-service": "^4.3.0",
     "uti": "1.10.6"
   },
   "devDependencies": {
     "chai": "^3.5.0",
     "cz-conventional-changelog": "^1.2.0",
     "istanbul": "^0.4.5",
-    "jsdoc": "^3.4.3",
     "mocha": "^3.2.0",
     "semantic-release": "^6.3.5",
     "jsdoc-to-markdown": "^2.0.1",
